refactor: Member 도메인 관련 피드백 반영(쿼리문 수정)
