+++ src/main/java/subway/member/ui/MemberController.java
@@ -4,10 +4,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.auth.domain.AuthenticationPrincipal;
-import subway.member.application.MemberService;
 import subway.member.domain.LoginMember;
 import subway.member.dto.MemberRequest;
 import subway.member.dto.MemberResponse;
+import subway.member.service.MemberService;
 
 import java.net.URI;
